# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Rest API for Trocon.",
        "title": "Trocon API",
        "contact": {},
        "version": "0.1"
    },
    "paths": {
        "/api/status": {
            "get": {
                "description": "Get server status.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Status"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/GetStatusAPI.GetUserResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/httputil.Error"
                            }
                        }
                    }
                }
            }
        },
        "/api/steam/user/:steamid": {
            "get": {
                "description": "Get server status.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Steam User"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Steam ID",
                        "name": "steamid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/GetSteamUserAPI.GetSteamUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httputil.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.Error"
                        }
                    }
                }
            }
        },
        "/api/steam/user/:steamid/trophy": {
            "get": {
                "description": "Get trophy list of steam user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Steam User"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Steam ID",
                        "name": "steamid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Steam appid list",
                        "name": "appid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/GetSteamUserTrophyAPI.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.Error"
                        }
                    }
                }
            }
        },
        "/api/steam/user/search": {
            "get": {
                "description": "Get steamid searched by keyword.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Steam User"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Steam ID or Steam vanity URL",
                        "name": "q",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/GetSteamUserSearchAPI.GetSteamUserSearchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httputil.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "Game.Game": {
            "type": "object",
            "properties": {
                "appId": {
                    "type": "integer"
                },
                "hasCommunityVisibleStats": {
                    "type": "boolean"
                },
                "headerImgUrl": {
                    "type": "string"
                },
                "iconImgUrl": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "playtime": {
                    "type": "integer"
                },
                "rtimeLastPlayed": {
                    "type": "integer"
                },
                "storeUrl": {
                    "type": "string"
                }
            }
        },
        "GetStatusAPI.GetUserResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "integer"
                }
            }
        },
        "GetSteamUserAPI.GetSteamUserResponse": {
            "type": "object",
            "properties": {
                "games": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Game.Game"
                    }
                },
                "statusCode": {
                    "type": "integer"
                },
                "user": {
                    "$ref": "#/definitions/steamworks.Player"
                }
            }
        },
        "GetSteamUserSearchAPI.GetSteamUserSearchResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "integer"
                },
                "steamid": {
                    "type": "string"
                }
            }
        },
        "GetSteamUserTrophyAPI.SuccessResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "integer"
                },
                "trophies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/GetSteamUserTrophyAPI.Trophy"
                    }
                }
            }
        },
        "GetSteamUserTrophyAPI.Trophy": {
            "type": "object",
            "properties": {
                "appId": {
                    "type": "integer"
                },
                "gameName": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                },
                "trophies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/steamworks.Achievement"
                    }
                }
            }
        },
        "httputil.Error": {
            "type": "object",
            "properties": {
                "errorCode": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        },
        "steamworks.Achievement": {
            "type": "object",
            "properties": {
                "achieved": {
                    "type": "integer"
                },
                "apiname": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "unlocktime": {
                    "type": "integer"
                }
            }
        },
        "steamworks.Player": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "avatarFull": {
                    "type": "string"
                },
                "avatarMedium": {
                    "type": "string"
                },
                "communityVisibilityState": {
                    "type": "integer"
                },
                "lastLogoff": {
                    "type": "integer"
                },
                "personaName": {
                    "type": "string"
                },
                "profileState": {
                    "type": "integer"
                },
                "profileUrl": {
                    "type": "string"
                },
                "steamId": {
                    "type": "string"
                }
            }
        }
    }
}